+++ test/exports.test.js
@@ -1,7 +1,11 @@
 import
 {
+	// Deprecated: `format()` was renamed to `formatNumber()`.
 	format,
+	formatNumber,
+	// Deprecated: `parse()` was renamed to `parseNumber()`.
 	parse,
+	parseNumber,
 	getNumberType,
 	isValidNumber,
 	findPhoneNumbers,
@@ -38,32 +42,37 @@ describe(`exports`, function()
 {
 	it(`should export ES6`, function()
 	{
-		parse('')
-		format('', 'US', 'National')
-		getNumberType('', 'RU')
-		isValidNumber('', 'US')
-		findPhoneNumbers('', 'US')
-		searchPhoneNumbers('', 'US')
-		new PhoneNumberSearch('')
-		new AsYouType('US').input('+')
+		// Deprecated: `parse()` was renamed to `parseNumber()`.
+		parse('+12133734253').should.deep.equal({ country: 'US', phone: '2133734253' })
+		parseNumber('+12133734253').should.deep.equal({ country: 'US', phone: '2133734253' })
+		// Deprecated: `format()` was renamed to `formatNumber()`.
+		format('2133734253', 'US', 'E.164').should.equal('+12133734253')
+		formatNumber('2133734253', 'US', 'E.164').should.equal('+12133734253')
+		getNumberType('2133734253', 'US').should.equal('FIXED_LINE_OR_MOBILE')
+		isValidNumber('+12133734253', 'US').should.equal(true)
+		findPhoneNumbers('+12133734253', 'US').should.deep.equal([{ country: 'US', phone: '2133734253', startsAt: 0, endsAt: 12 }])
+		searchPhoneNumbers('+12133734253', 'US')[Symbol.iterator]().next.should.be.a('function')
+		new PhoneNumberSearch('+12133734253', undefined).find.should.be.a('function')
+		new AsYouType('US').input('+12133734253').should.equal('+1 213 373 4253')
 
-		DIGIT_PLACEHOLDER.length
+		DIGIT_PLACEHOLDER.should.equal('x')
 		Object.keys(DIGITS).length.should.be.above(0)
 
 		// `getPhoneCode` name is deprecated.
-		getPhoneCode('KZ')
-		getCountryCallingCode('KZ')
+		getPhoneCode('KZ').should.equal('7')
+		getCountryCallingCode('KZ').should.equal('7')
 
 		new Metadata({ countries: {} })
-		parseRFC3966.should.be.a('function')
-		formatRFC3966.should.be.a('function')
+		parseRFC3966('tel:+12133734253').should.deep.equal({ number: '+12133734253' })
+		formatRFC3966({ number: '+12133734253' }).should.equal('tel:+12133734253')
 	})
 
+	// Deprecated exports: remove in `2.0.0`.
 	it(`should export ES6 custom functions`, function()
 	{
-		parseCustom('', metadata)
-		formatCustom('', 'US', 'National', metadata)
-		getNumberTypeCustom('', 'RU', metadata)
+		parseCustom('+12133734253', metadata).should.deep.equal({ country: 'US', phone: '2133734253' })
+		formatCustom('2133734253', 'US', 'E.164', metadata).should.equal('+12133734253')
+		getNumberTypeCustom('2133734253', 'US', metadata).should.equal('FIXED_LINE_OR_MOBILE')
 		isValidNumberCustom('', 'US', metadata)
 		findPhoneNumbers('', 'US', metadata)
 		searchPhoneNumbers('', 'US', metadata)
@@ -77,40 +86,51 @@ describe(`exports`, function()
 	{
 		const Library = require('../index.common')
 
-		Library.parse('')
-		Library.format('', 'US', 'National')
-		Library.getNumberType('', 'RU')
-		Library.isValidNumber('', 'US')
-		Library.findPhoneNumbers('', 'US')
-		Library.searchPhoneNumbers('', 'US')
-		new Library.PhoneNumberSearch('')
-		new Library.AsYouType('US').input('+')
+		// Deprecated: `parse()` was renamed to `parseNumber()`.
+		Library.parse('+12133734253').should.deep.equal({ country: 'US', phone: '2133734253' })
+		Library.parseNumber('+12133734253').should.deep.equal({ country: 'US', phone: '2133734253' })
+		// Deprecated: `format()` was renamed to `formatNumber()`.
+		Library.format('2133734253', 'US', 'E.164').should.equal('+12133734253')
+		Library.formatNumber('2133734253', 'US', 'E.164').should.equal('+12133734253')
+		Library.getNumberType('2133734253', 'US').should.equal('FIXED_LINE_OR_MOBILE')
+		Library.isValidNumber('+12133734253', 'US').should.equal(true)
+		Library.findPhoneNumbers('+12133734253', 'US').should.deep.equal([{ country: 'US', phone: '2133734253', startsAt: 0, endsAt: 12 }])
+		Library.searchPhoneNumbers('+12133734253', 'US')[Symbol.iterator]().next.should.be.a('function')
+		new Library.PhoneNumberSearch('+12133734253', undefined).find.should.be.a('function')
+		new Library.AsYouType('US').input('+12133734253').should.equal('+1 213 373 4253')
 
-		Library.DIGIT_PLACEHOLDER.length
+		Library.DIGIT_PLACEHOLDER.should.equal('x')
 		Object.keys(Library.DIGITS).length.should.be.above(0)
 
 		// `getPhoneCode` name is deprecated.
-		Library.getPhoneCode('KZ')
-		Library.getCountryCallingCode('KZ')
+		Library.getPhoneCode('KZ').should.equal('7')
+		Library.getCountryCallingCode('KZ').should.equal('7')
+
+		Library.parseRFC3966('tel:+12133734253').should.deep.equal({ number: '+12133734253' })
+		Library.formatRFC3966({ number: '+12133734253' }).should.equal('tel:+12133734253')
 	})
 
 	it(`should export CommonJS custom functions`, function()
 	{
 		const Library = require('../custom')
 
-		Library.parse('', metadata)
-		Library.format('', 'US', 'National', metadata)
-		Library.getNumberType('', 'RU', metadata)
-		Library.isValidNumber('', 'US', metadata)
-		Library.findPhoneNumbers('', 'US', metadata)
-		Library.searchPhoneNumbers('', 'US', metadata)
-		new Library.PhoneNumberSearch('', undefined, metadata)
+		// Deprecated: `parse()` was renamed to `parseNumber()`.
+		Library.parse('+12133734253', metadata).should.deep.equal({ country: 'US', phone: '2133734253' })
+		Library.parseNumber('+12133734253', metadata).should.deep.equal({ country: 'US', phone: '2133734253' })
+		// Deprecated: `format()` was renamed to `formatNumber()`.
+		Library.format('2133734253', 'US', 'E.164', metadata).should.equal('+12133734253')
+		Library.formatNumber('2133734253', 'US', 'E.164', metadata).should.equal('+12133734253')
+		Library.getNumberType('2133734253', 'US', metadata).should.equal('FIXED_LINE_OR_MOBILE')
+		Library.isValidNumber('+12133734253', 'US', metadata).should.equal(true)
+		Library.findPhoneNumbers('+12133734253', 'US', metadata).should.deep.equal([{ country: 'US', phone: '2133734253', startsAt: 0, endsAt: 12 }])
+		Library.searchPhoneNumbers('+12133734253', 'US', metadata)[Symbol.iterator]().next.should.be.a('function')
+		new Library.PhoneNumberSearch('+12133734253', undefined, metadata).find.should.be.a('function')
 		// `getPhoneCode` name is deprecated.
-		Library.getPhoneCode('KZ', metadata)
-		Library.getCountryCallingCode('KZ', metadata)
-
+		Library.getPhoneCode('KZ', metadata).should.equal('7')
+		Library.getCountryCallingCode('KZ', metadata).should.equal('7')
+		new Library.AsYouType('US', metadata).input('+12133734253').should.equal('+1 213 373 4253')
 		new Library.Metadata({ countries: {} })
-		Library.parseRFC3966.should.be.a('function')
-		Library.formatRFC3966.should.be.a('function')
+		Library.parseRFC3966('tel:+12133734253').should.deep.equal({ number: '+12133734253' })
+		Library.formatRFC3966({ number: '+12133734253' }).should.equal('tel:+12133734253')
 	})
 })